+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## develop
 - add ElectricLoad.blended_doe_reference_names & blended_doe_reference_percents
+- add ElectricLoad.monthly_totals_kwh builtin profile scaling
 
 ## v0.10.0
 - add modeling capability for tiered rates (energy, TOU demand, and monthly demand charges)